+++ assets/js/util/utils.ts
@@ -345,14 +345,6 @@ export function div(): HTMLDivElement {
   return document.createElement("div") as HTMLDivElement;
 }
 
-// FIXME: These should be replaced with config based deciding of "main" sample
-export function isNonMainSample(sample: Sample): boolean {
-  const isNotMainSample =
-    sample.sampleType == null ||
-    ["proband", "tumor"].includes(sample.sampleType);
-  return !isNotMainSample;
-}
-
 export function getMainSample(samples: Sample[]): Sample {
   const mainSample = samples.find((s) =>
     ["proband", "tumor"].includes(s.sampleType),